requirements.txt
# requirements.txt
yt-dlp
Pillow
requests
librosa
PyQt5
# (no pyinstaller here)

